edit draw_map to center rendered object and added few key hook

+++ srcs/draw_line.c
@@ -6,7 +6,7 @@
 /*   By: hferraud <hferraud@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 07:55:53 by hferraud          #+#    #+#             */
-/*   Updated: 2022/12/04 01:37:26 by hferraud         ###   ########lyon.fr   */
+/*   Updated: 2022/12/05 03:33:07 by hferraud         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	draw_hrz(t_vec_3d p_start, t_vec_3d p_end, t_data *imgdata)
 	double	coef;
 	double	exact_y;
 
-	coef = (double)(p_end.y - p_start.y) / (double)(p_end.x - p_start.x);
+	coef = (p_end.y - p_start.y) / (p_end.x - p_start.x);
 	put_pixel(imgdata, p_start.x, p_start.y, 0xffffff);
 	exact_y = p_start.y;
 	while (p_start.x <= p_end.x)
@@ -35,7 +35,7 @@ void	draw_vrt(t_vec_3d p_start, t_vec_3d p_end, t_data *imgdata)
 	double	coef;
 	double	exact_x;
 
-	coef = (double)(p_end.x - p_start.x) / (double)(p_end.y - p_start.y);
+	coef = (p_end.x - p_start.x) / (p_end.y - p_start.y);
 	put_pixel(imgdata, p_start.x, p_start.y, 0xffffff);
 	exact_x = p_start.x;
 	while (p_start.y <= p_end.y)
@@ -53,7 +53,7 @@ void	draw_line(t_vec_3d p_start, t_vec_3d p_end, t_data *imgdata)
 
 	if (p_start.x > p_end.x)
 		return (draw_line(p_end, p_start, imgdata));
-	coef = (double)(p_end.y - p_start.y) / (double)(p_end.x - p_start.x);
+	coef = (p_end.y - p_start.y) / (p_end.x - p_start.x);
 	if (coef > -0.5 && coef < 0.5)
 		return (draw_hrz(p_start, p_end, imgdata));
 	if (p_start.y > p_end.y)
